chore: refactor CheckpointViewButton into seperate component
